CRUD cupom

+++ src/core/Ecommerce.Domain/Entity/Coupon.cs
@@ -3,7 +3,7 @@
 namespace Ecommerce.Domain.Entity;
 
 public record Coupon
-    (Guid Id, DateTime CreatedIn, DateTime UpdatedIn, string Code, int DiscountPercentage, DateTime? ValidUntil) 
+    (Guid Id, DateTime CreatedIn, DateTime UpdatedIn, string Code, float DiscountPercentage, DateTime? ValidUntil) 
         : BaseEntity(Id, CreatedIn, UpdatedIn)
 {
     public ICollection<Product> Products { get; } = default!;
